fsck.overlay: fix two compile problems
Fix following two compile warnings and one ld error with GLIBC 2.28 and
GCC 8.3.0.

In file included from lib.c:34:
lib.c: In function ‘scan_dir’:
common.h:32:15: warning: format ‘%jd’ expects argument of type ‘intmax_t’, but argument 4 has type ‘__off_t’ {aka ‘long int’} [-Wformat=]
 #define _(x)  (x)
               ^~~
lib.c:227:15: note: in expansion of macro ‘_’
   print_debug(_("Scan:%-3s %2d %7jd   %-40s %-20s\n"),
               ^
gcc -Wall -g -c check.c
check.c: In function ‘ovl_create_whiteout’:
check.c:173:55: warning: implicit declaration of function ‘makedev’ [-Wimplicit-function-declaration]
  if (mknodat(dirfd, pathname, S_IFCHR | WHITEOUT_MOD, makedev(0, 0))) {
                                                       ^~~~~~~
gcc -lm fsck.o common.o lib.o check.o mount.o path.o overlayfs.o -o fsck.overlay
/usr/bin/ld: check.o: in function `ovl_create_whiteout':

Reported-by: Chengguang Xu <cgxu519@gmx.com>

+++ lib.c
@@ -224,7 +224,7 @@ int scan_dir(struct scan_ctx *sctx, struct scan_operations *sop)
 		/* Fillup base context */
 		scan_entry_init(sctx, ftsent);
 
-		print_debug(_("Scan:%-3s %2d %7jd   %-40s %-20s\n"),
+		print_debug(_("Scan:%-3s %2d %7lld   %-40s %-20s\n"),
 			      (ftsent->fts_info == FTS_D) ? "d" :
 			      (ftsent->fts_info == FTS_DNR) ? "dnr" :
 			      (ftsent->fts_info == FTS_DP) ? "dp" :
@@ -233,7 +233,8 @@ int scan_dir(struct scan_ctx *sctx, struct scan_operations *sop)
 			      (ftsent->fts_info == FTS_SL) ? "sl" :
 			      (ftsent->fts_info == FTS_SLNONE) ? "sln" :
 			      (ftsent->fts_info == FTS_DEFAULT) ? "df" : "???",
-			      ftsent->fts_level, ftsent->fts_statp->st_size,
+			      ftsent->fts_level,
+			      (long long)ftsent->fts_statp->st_size,
 			      ftsent->fts_path, sctx->layer->path);
 
 		switch (ftsent->fts_info) {
